Added new edit pane and buttons now disable rather than disappear.
Fairly major interface changes, buttons now tend to disable rather than not
be created at all when they are not relevant. This fixes the issue of the
buttons suddenly changing size and position.

A new edit panel has also been added, which is where the delete option is
now located. This panel should be used for useful tools such as undo and
delete as well later on.

Other minor changes include updating the labels on buttons so that it is
more consistent (capitalization, and no longer random periods on some and
not others.)