Couple of usability fixes. Save now updates the source panel.
Saving a file will now load the source code into the source viewer after
writing to disk.

Also, the default name for a new files root node is no longer has spaces in
it, as the file loader / parser cannot deal with spaces. Meant that the
default names were unreadable by the thing that created them...